refactor: add super classes
