+++ ST/Pavel.Pleshkov/Project/src/components/index.js
@@ -1,12 +1,6 @@
-import Basket from './basket';
-import Catalog from './catalog';
+import { Basket, Catalog } from './LISTS';
 
 export default () => {
-  let catalogUrl = 'https://raw.githubusercontent.com/kellolo/static/master/JSON/catalog.json';
-  let basketgUrl = 'https://raw.githubusercontent.com/kellolo/static/master/JSON/basket.json';
-
-  let basket = new Basket('#basket', basketgUrl);
-  let catalog = new Catalog('#catalog', catalogUrl, basket);
-  // basket.init();
-  // catalog.init(basket); 
+    let basket = new Basket();
+    let catalog = new Catalog(basket);
 }
